+++ src/Auth/actions.js
@@ -1,0 +1,4 @@
+export const AUTH_USER = 'AUTH_USER'
+export const AUTH_HOUSE = 'AUTH_HOUSE'
+export const AUTH_LOGOUT = 'AUTH_LOGOUT'
+export const REFRESH_TOKEN = 'REFRESH_TOKEN'
