+++ src/Store/actions/actionTypes.js
@@ -1,8 +1,3 @@
-export const AUTH_USER = 'AUTH_USER'
-export const AUTH_HOUSE = 'AUTH_HOUSE'
-export const AUTH_LOGOUT = 'AUTH_LOGOUT'
-export const REFRESH_TOKEN = 'REFRESH_TOKEN'
-
 export const START_LOADING = 'START_LOADING'
 export const STOP_LOADING = 'STOP_LOADING'
